ltrace: add new package at version 0.7.3
ltrace intercepts and records dynamic library calls which are called
by an executed process and the signals received by that process.
It can also intercept and print the system calls executed by the program.
ltrace was originally authored by Juan Cespedes.

It is to library calls what strace is for syscalls.

OpenWRT: currently does not work very well. It does not work with stripped
binaries (OpenWRT default) and does not work with ar71xx (mips).

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>